+++ _PostInstall.ps1
@@ -19,47 +19,6 @@ Start-Sleep -Seconds 2
 Write-Host .
 Write-Host .
 Write-Host ---------------------------------------------------------------------------------
-Write-Host -------------------------------CREATE FOLDERS------------------------------------
-Write-Host ---------------------------------------------------------------------------------
-New-Item -Path "C:\" -Name "TEMP" -ItemType "directory" 
-New-Item -Path "C:\" -Name "!DOWNLOADS" -ItemType "directory" 
-New-Item -Path "C:\" -Name "!APPS" -ItemType "directory" 
-New-Item -Path "C:\" -Name "!GAMES" -ItemType "directory" 
-New-Item -Path "C:\" -Name "!MYCOR" -ItemType "directory" 
-Write-Host .
-Write-Host .
-Start-Sleep -Seconds 2
-Write-Host .
-Write-Host .
-Write-Host ---------------------------------------------------------------------------------
-Write-Host -------------------------------INSTALL FONTS-------------------------------------
-Write-Host ---------------------------------------------------------------------------------
-Copy-Item -Path "C:\Windows\hack.ttf" -Destination "C:\Windows\Fonts"
-Write-Host .
-Write-Host .
-Start-Sleep -Seconds 2
-Write-Host .
-Write-Host .
-Write-Host ---------------------------------------------------------------------------------
-Write-Host --------------------------EDIT ENVIROMENTAL VARIABLES----------------------------
-Write-Host ---------------------------------------------------------------------------------
-[Environment]::SetEnvironmentVariable("Path", $env:PATH + ";C:\BATCH;C:\PortableApps\VSCode\bin;C:\PortableApps\Git\bin;C:\PortableApps\nvim\bin", [System.EnvironmentVariableTarget]::User)
-[Environment]::SetEnvironmentVariable("Path", $env:PATH + ";C:\BATCH;C:\PortableApps\VSCode\bin;C:\PortableApps\Git\bin;C:\PortableApps\nvim\bin",  [System.EnvironmentVariableTarget]::Machine)
-Write-Host .
-Write-Host .
-Start-Sleep -Seconds 2
-Write-Host .
-Write-Host .
-Write-Host ---------------------------------------------------------------------------------
-Write-Host -------------------------------TOOLBAR LINKS-------------------------------------
-Write-Host ---------------------------------------------------------------------------------
-Copy-Item -Path "C:\BATCH\Links\*" -Destination "C:\Users\$env:USERNAME\Favorites\Links" -Recurse
-Write-Host .
-Write-Host .
-Start-Sleep -Seconds 2
-Write-Host .
-Write-Host .
-Write-Host ---------------------------------------------------------------------------------
 Write-Host -------------------------------EDITING SWAP--------------------------------------
 Write-Host ---------------------------------------------------------------------------------
 # Uncheck automatically set pagefile
@@ -71,15 +30,6 @@ $pagefileset = Get-WmiObject Win32_pagefilesetting
 $pagefileset.InitialSize = 8000
 $pagefileset.MaximumSize = 8000
 $pagefileset.Put() | Out-Null
-Write-Host .
-Write-Host .
-Start-Sleep -Seconds 2
-Write-Host .
-Write-Host .
-Write-Host ---------------------------------------------------------------------------------
-Write-Host -------------------------------DISABLE SERVICES----------------------------------
-Write-Host ---------------------------------------------------------------------------------
-Stop-Service -Name "ShellHWDetection" -Force
 Write-Host .
 Write-Host .
 Start-Sleep -Seconds 2
@@ -115,24 +65,58 @@ Write-Host .
 Write-Host ---------------------------------------------------------------------------------
 Write-Host -------------------------------UNINSTALL BLOAT-----------------------------------
 Write-Host ---------------------------------------------------------------------------------
-Write-Host "Uninstall bloat? (y/n)"
-$uninstall_bloat = Read-Host -Prompt "You options is"
-if ($uninstall_bloat  -like 'y') {
-    C:\_DebloatScript.ps1
-}
-if ($uninstall_bloat  -like 'n') {
-    Start-Sleep -Seconds 1
-}
-Start-Sleep -Seconds 2
+Write-Host .
+Write-Host .
+Get-AppxPackage -allusers Disney.37853FC22B2CE | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.549981C3F5F10 | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.BingNews | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.BingWeather | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.GetHelp | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Getstarted | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Microsoft3DViewer | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MicrosoftOfficeHub | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MicrosoftSolitaireCollection | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MicrosoftStickyNotes | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MixedReality.Portal | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.MSPaint | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Office.OneNote | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.OneDriveSync | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.People | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.PowerAutomateDesktop | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.SkypeApp | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Todos | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.Wallet | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsAlarms | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsCamera | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsFeedbackHub | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsMaps | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.WindowsSoundRecorder | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.YourPhone | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.ZuneMusic | Remove-AppxPackage
+Get-AppxPackage -allusers Microsoft.ZuneVideo | Remove-AppxPackage
+Get-AppxPackage -allusers MicrosoftTeams | Remove-AppxPackage
+Get-AppxPackage -allusers MicrosoftWindows.Client.WebExperience | Remove-AppxPackage
+Get-AppxPackage -allusers SpotifyAB.SpotifyMusic | Remove-AppxPackage
+Get-AppxPackage -allusers microsoft.windowscommunicationsapps | Remove-AppxPackage
+Get-AppxPackage -allusers MicrosoftTeams* | Remove-AppxPackage
+# Remove all xbox stuff
+Get-ProvisionedAppxPackage -Online | `
+Where-Object { $_.PackageName -match "xbox" } | `
+ForEach-Object { Remove-ProvisionedAppxPackage -Online -AllUsers -PackageName $_.PackageName }
+C:\Windows\SysWOW64\OneDriveSetup.exe -uninstall
+Set-Location -Path "C:\Program Files (x86)\Microsoft\Edge\Application\9*\Installer" 
+./setup.exe --uninstall --force-uninstall --system-level
+Get-WindowsPackage -Online | Where PackageName -like *Hello-Face* | Remove-WindowsPackage -Online -NoRestart
+Get-WindowsPackage -Online | Where PackageName -like *QuickAssist* | Remove-WindowsPackage -Online -NoRestart
 Write-Host .
 Write-Host .
 Write-Host ---------------------------------------------------------------------------------
-Write-Host -------------------------------SETUP MY APPS-------------------------------------
+Write-Host -------------------------------SETUP MY SETTINGS-------------------------------------
 Write-Host ---------------------------------------------------------------------------------
-Write-Host "Setup my apps? (y/n)"
+Write-Host "Setup my setting? (y/n)"
 $setup_my_apps = Read-Host -Prompt "You options is"
 if ($setup_my_apps -like 'y') {
-    C:\_AppSetup.ps1
+    C:\_PersonalSetup.ps1
 }
 if ($setup_my_apps -like 'n') {
     Start-Sleep -Seconds 1
@@ -177,23 +161,6 @@ Start-Sleep -Seconds 15
 Write-Host .
 Write-Host .
 Write-Host ---------------------------------------------------------------------------------
-Write-Host ----------------------------CAPS LIKE CHANGE LAYOUT------------------------------
-Write-Host ---------------------------------------------------------------------------------
-C:\Windows\engcaps\setup.exe 
-Start-Sleep -Seconds 15
-Stop-Process -name setup -force
-C:\Windows\ruscaps\setup.exe 
-Start-Sleep -Seconds 15
-Stop-Process -name setup -force
-C:\Windows\ukrcaps\setup.exe 
-Start-Sleep -Seconds 15
-Stop-Process -name setup -force
-Write-Host .
-Write-Host .
-Start-Sleep -Seconds 2
-Write-Host .
-Write-Host .
-Write-Host ---------------------------------------------------------------------------------
 Write-Host ----------------------------DISABLING NETWORK DISCOVERY--------------------------
 Write-Host ---------------------------------------------------------------------------------
 netsh advfirewall firewall set rule group="Network Discovery" new enable=No
@@ -229,12 +196,10 @@ Remove-Item C:\Windows\ukrcaps -Recurse
 Remove-Item C:\Windows\_BitsumHighestPerformance.pow
 Remove-Item C:\Windows\_C++.exe
 Remove-Item C:\Windows\_Registry.reg
-Remove-Item C:\Windows\_WinRAR.reg
-Remove-Item C:\Windows\_Telegram.reg
-remove-item C:\Windows\hack.ttf
-remove-item C:\Windows\rarreg.key
+Remove-Item C:\Windows\_PersonalRegistry.reg
+Remove-item C:\Windows\hack.ttf
+Remove-item C:\Windows\rarreg.key
 Remove-Item C:\_PostInstall.cmd
-Remove-Item C:\_DebloatScript.ps1
-Remove-Item C:\_AppSetup.ps1
+Remove-Item C:\_PersonalSetup.ps1
 Restart-Computer
 Remove-Item C:\_PostInstall.ps1